+++ site/docs/10-physics/examples/physics.ts
@@ -1,3 +1,10 @@
+import * as ex from 'excalibur';
+
+const game = new ex.Engine({
+  canvasElementId: 'preview-canvas',
+  displayMode: ex.DisplayMode.FillContainer,
+});
+
 const box = new ex.Actor({
     pos: ex.vec(game.halfDrawWidth, -100),
     width: 50,